backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.8
sqlalchemy==2.0.36
opencv-python-headless==4.10.0.84
pillow==11.0.0
reportlab==4.2.4
ultralytics==8.3.53
